pyproject.toml
[project]
name = "douglas"
version = "0.1.0"
description = "AI-assisted development loop orchestrator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Dicky Moore" }]
dependencies = ["typer>=0.9", "pyyaml>=6.0"]

[project.scripts]
douglas = "douglas.cli:app"

[project.optional-dependencies]
dev = ["pytest", "ruff", "black", "isort", "mypy", "types-PyYAML"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["douglas*"]

[tool.isort]
profile = "black"
